Month 13/step 2

+++ src/controllers/card-controller.ts
@@ -1,13 +1,13 @@
 import { Request, Response, NextFunction } from 'express';
-import { ITestRequest } from '../middleware/middleware';
+import { IUserRequest } from '../utils/type-user-request';
 import Cards from '../models/card';
 import AppError from '../errors/custom-errors';
 
 const getCards = (req: Request, res: Response, next: NextFunction) => Cards.find({})
   .then((cards) => res.send({ data: cards }))
-  .catch(() => next(AppError.serverError('Server error')));
+  .catch((err) => next(err));
 
-const createCard = (req: ITestRequest, res: Response, next: NextFunction) => {
+const createCard = (req: IUserRequest, res: Response, next: NextFunction) => {
   const { name, link } = req.body;
   const owner = req.user?._id;
   return Cards.create({ name, link, owner })
@@ -18,11 +18,11 @@ const createCard = (req: ITestRequest, res: Response, next: NextFunction) => {
       if (err instanceof Error && err.name === 'ValidationError') {
         return next(AppError.badRequest('Incorrect data'));
       }
-      return next(AppError.serverError('Server error'));
+      return next(err);
     });
 };
 
-const putLikeCard = (req: ITestRequest, res: Response, next: NextFunction) => {
+const putLikeCard = (req: IUserRequest, res: Response, next: NextFunction) => {
   const _id = req.user?._id;
   const { cardId } = req.params;
   return Cards.findByIdAndUpdate(
@@ -40,12 +40,12 @@ const putLikeCard = (req: ITestRequest, res: Response, next: NextFunction) => {
       if (err instanceof Error && err.name === 'CastError') {
         return next(AppError.badRequest('Incorrect data'));
       }
-      return next(AppError.serverError('Server error'));
+      return next(err);
     });
 };
 
 const deleteLikeCard = (
-  req: ITestRequest,
+  req: IUserRequest,
   res: Response,
   next: NextFunction,
 ) => {
@@ -66,24 +66,32 @@ const deleteLikeCard = (
       if (err instanceof Error && err.name === 'CastError') {
         return next(AppError.badRequest('Incorrect data'));
       }
-      return next(AppError.serverError('Server error'));
+      return next(err);
     });
 };
 
-const deleteCard = (req: ITestRequest, res: Response, next: NextFunction) => {
+const deleteCard = (req: IUserRequest, res: Response, next: NextFunction) => {
   const { cardId } = req.params;
-  return Cards.findByIdAndDelete(cardId)
-    .then((card) => {
-      if (!card) {
+  const owner = req.user?._id;
+  Cards.findById(cardId)
+    .then((cardDelete) => {
+      if (!cardDelete) {
         throw AppError.notFound('Card not found');
       }
-      return res.send({ data: card });
+      if (cardDelete.owner.toString() === owner) {
+        return Cards.findByIdAndDelete(cardId)
+          .then((card) => res.send({ data: card }))
+          .catch((err) => next(err));
+      }
+      return next(
+        AppError.unathorized('You do not have the right to delete card'),
+      );
     })
     .catch((err) => {
       if (err instanceof Error && err.name === 'CastError') {
         return next(AppError.badRequest('Incorrect data'));
       }
-      return next(AppError.serverError('Server error'));
+      return next(err);
     });
 };
